requirements.txt
Django>=5.2.7
djangorestframework>=3.15
gunicorn==23.0.0
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
whitenoise>=6.6
dj-database-url>=1.0
packaging==25.0
